Module reorganization, documentation, and common methods.
