+++ package.json
@@ -43,10 +43,12 @@
   },
   "devDependencies": {
     "@types/jest": "^29.5.5",
+    "@types/luxon": "^3.3.2",
     "@types/qs": "^6.9.8",
     "eslint-config-ethang": "^11.0.0",
     "eslint-plugin-jest": "^27.4.0",
     "jest": "^29.7.0",
+    "luxon": "^3.4.3",
     "ts-jest": "^29.1.1",
     "tsup": "^7.2.0",
     "typescript": "^5.2.2",
