+++ src/index.ts
@@ -1,8 +1,10 @@
+// Common
+export { dateInWhereFormat } from './common';
 // Contacts
-export { getContactLink } from './contacts';
+export { getContactLink } from './accounting/contacts';
 // Invoices
-export { filterInvoiceLineItems, getInvoiceLink } from './invoices';
+export { filterInvoiceLineItems, getInvoiceLink } from './accounting/invoices';
 // Journals
-export { getManualJournalLink } from './journals';
+export { getManualJournalLink } from './accounting/journals';
 // Projects
 export { generateProjectAmountUSD } from './projects';
